A follow up on #2 that unifies what we put in id-map
This is not intended to replace #2 but to present an alternative that we can discuss.

When we have an ident such as `[:id "whatever"]` then `entity-triples` used by the map form of tx-data would insert just "whatever" into the id-map `ids` upon creating the entity. That is contrary to the code just added in 9ea60eb, which puts the whole ident in there.

We want to unify both so that references will be resolved correctly no matter what (map x triplets) form is used.

This commits bows to the wisdom of entity-triples and puts only the ident's value into the id-map.

BEWARE: An ident's value can be _anything_. Thus it puts us at risk when we do the `(ids %)` lookup at L208 that we will replace some E/A/V value that was not meant as a reference.